requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.2
cryptography==46.0.7
Deprecated==1.3.1
fastapi==0.135.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
libsql-experimental==0.0.55
limits==5.8.0
markdown-it-py==4.0.0
mcp==1.27.0
mdurl==0.1.2
packaging==26.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
referencing==0.37.0
rich==14.3.4
rpds-py==0.30.0
shellingham==1.5.4
slowapi==0.1.9
sse-starlette==3.3.4
starlette==1.0.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
wrapt==2.1.2
